+++ src/flash/nand/tcl.c
@@ -341,7 +341,7 @@ COMMAND_HANDLER(handle_nand_verify_command)
 
 COMMAND_HANDLER(handle_nand_dump_command)
 {
-	int filesize;
+	size_t filesize;
 	struct nand_device *nand = NULL;
 	struct nand_fileio_state s;
 	int retval = CALL_COMMAND_HANDLER(nand_fileio_parse_args,
@@ -374,8 +374,8 @@ COMMAND_HANDLER(handle_nand_dump_command)
 		return retval;
 
 	if (nand_fileio_finish(&s) == ERROR_OK) {
-		command_print(CMD_CTX, "dumped %ld bytes in %fs (%0.3f KiB/s)",
-			(long)filesize, duration_elapsed(&s.bench),
+		command_print(CMD_CTX, "dumped %zu bytes in %fs (%0.3f KiB/s)",
+			filesize, duration_elapsed(&s.bench),
 			duration_kbps(&s.bench, filesize));
 	}
 	return ERROR_OK;
